refactor: move to handlers directory
