+++ wdl-analysis/src/handlers.rs
@@ -1,0 +1,5 @@
+//! Language server protocol handlers.
+
+mod goto_definition;
+
+pub use goto_definition::*;
